+++ app/src/main/java/org/greenstand/android/TreeTracker/view/DepthButtonColors.kt
@@ -1,0 +1,61 @@
+/*
+ * Copyright 2023 Treetracker
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.greenstand.android.TreeTracker.view
+
+import androidx.compose.material.ButtonColors
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Immutable
+import androidx.compose.runtime.State
+import androidx.compose.runtime.rememberUpdatedState
+import androidx.compose.ui.graphics.Color
+
+@Immutable
+class DepthButtonColors(
+    val shadowColor: Color,
+    val color: Color,
+    val disabledShadowColor: Color,
+    val disabledColor: Color,
+) : ButtonColors {
+    @Composable
+    override fun backgroundColor(enabled: Boolean): State<Color> =
+        rememberUpdatedState(if (enabled) shadowColor else disabledShadowColor)
+
+    @Composable
+    override fun contentColor(enabled: Boolean): State<Color> =
+        rememberUpdatedState(if (enabled) color else disabledColor)
+
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other == null || this::class != other::class) return false
+
+        other as DepthButtonColors
+
+        if (shadowColor != other.shadowColor) return false
+        if (color != other.color) return false
+        if (disabledShadowColor != other.disabledShadowColor) return false
+        if (disabledColor != other.disabledColor) return false
+
+        return true
+    }
+
+    override fun hashCode(): Int {
+        var result = shadowColor.hashCode()
+        result = 31 * result + color.hashCode()
+        result = 31 * result + disabledShadowColor.hashCode()
+        result = 31 * result + disabledColor.hashCode()
+        return result
+    }
+}
